requirements.txt
fastapi
uvicorn
diffusers
transformers
accelerate
invisible_watermark
torch
pillow
requests 
fastapi 
uvicorn
streamlit